+++ eqt/ui/UIFormWidget.py
@@ -2,6 +2,8 @@ from PySide2 import QtWidgets
 
 from .UISliderWidget import UISliderWidget
 
+from warnings import warn
+
 
 class UIFormWidget:
     '''
@@ -463,6 +465,11 @@ class UIFormWidget:
         To later restore the states, use `restoreAllSavedWidgetStates()`.
         '''
         self.widget_states = self.getAllWidgetStates()
+    
+    def getWidgetStates(self):
+        '''Deprecated. Use `getSavedWidgetStates`.'''
+        warn('The method `getWidgetStates` is deprecated, use `getSavedWidgetStates`.', DeprecationWarning, stacklevel=2)
+        return self.getSavedWidgetStates()
 
     def getSavedWidgetStates(self):
         '''Returns the saved widget states.'''
@@ -621,6 +628,10 @@ class FormDockWidget(QtWidgets.QDockWidget):
         To later restore the states, use `restoreAllSavedWidgetStates()`.
         '''
         self.widget().saveAllWidgetStates()
+
+    def getWidgetStates(self):
+        '''Deprecated. Use `getSavedWidgetStates`.'''
+        return self.widget().getWidgetStates()
 
     def getSavedWidgetStates(self):
         '''Returns the saved widget states.'''